Move `PowerDistributingActor` to the `microgrid` module
