+++ src/frequenz/sdk/microgrid/_power_distributing/_component_managers/_pv_inverter_manager/_pv_inverter_manager.py
@@ -18,8 +18,8 @@ from frequenz.client.microgrid import (
 from typing_extensions import override
 
 from ....._internal._math import is_close_to_zero
-from .....microgrid import connection_manager
 from .....timeseries import Power
+from .... import connection_manager
 from ..._component_pool_status_tracker import ComponentPoolStatusTracker
 from ..._component_status import ComponentPoolStatus, PVInverterStatusTracker
 from ...request import Request
