+++ src/frequenz/sdk/microgrid/_power_distributing/_component_pool_status_tracker.py
