+++ src/frequenz/sdk/microgrid/_power_distributing/_component_status/_blocking_status.py
