+++ src/frequenz/sdk/microgrid/_power_distributing/_component_status/_pv_inverter_status_tracker.py
@@ -1,8 +1,7 @@
 # License: MIT
 # Copyright © 2024 Frequenz Energy-as-a-Service GmbH
 
-"""Background service that tracks the status of an EV charger."""
-
+"""Background service that tracks the status of a PV inverter."""
 
 import asyncio
 import logging
@@ -10,16 +9,11 @@ from datetime import datetime, timedelta, timezone
 
 from frequenz.channels import Receiver, Sender, select, selected_from
 from frequenz.channels.timer import SkipMissedAndDrift, Timer
-from frequenz.client.microgrid import (
-    EVChargerCableState,
-    EVChargerComponentState,
-    EVChargerData,
-)
+from frequenz.client.microgrid import InverterComponentState, InverterData
 from typing_extensions import override
 
-from frequenz.sdk.microgrid import connection_manager
-
-from ..._background_service import BackgroundService
+from ....actor._background_service import BackgroundService
+from ... import connection_manager
 from ._blocking_status import BlockingStatus
 from ._component_status import (
     ComponentStatus,
@@ -31,11 +25,13 @@ from ._component_status import (
 _logger = logging.getLogger(__name__)
 
 
-class EVChargerStatusTracker(ComponentStatusTracker, BackgroundService):
-    """Status tracker for EV chargers.
+class PVInverterStatusTracker(ComponentStatusTracker, BackgroundService):
+    """Status tracker for PV inverters.
 
-    It reports an EV charger as `WORKING` when an EV is connected to it,
-    and power can be allocated to it, and `NOT_WORKING` otherwise.
+    It reports a PV inverter as `WORKING` or `NOT_WORKING` based on
+    the status in the received component data from the microgrid API.
+    When no data is received for a specific duration, the component is
+    marked as `NOT_WORKING`.
 
     If it receives a power assignment failure from the PowerDistributor,
     when the component is expected to be `WORKING`, it is marked as
@@ -55,16 +51,17 @@ class EVChargerStatusTracker(ComponentStatusTracker, BackgroundService):
         """Initialize this instance.
 
         Args:
-            component_id: ID of the EV charger to monitor the status of.
+            component_id: ID of the PV inverter to monitor the status of.
             max_data_age: max duration to wait for, before marking a component as
                 NOT_WORKING, unless new data arrives.
-            max_blocking_duration: duration for which the component status should be
-                UNCERTAIN if a request to the component failed unexpectedly.
-            status_sender: Channel sender to send status updates to.
-            set_power_result_receiver: Receiver to fetch PowerDistributor responses
-                from, to get the status of the most recent request made for an EV
-                Charger.
+            max_blocking_duration: max duration to wait for, before marking a component
+                as BLOCKING, unless new data arrives.
+            status_sender: Sender to send the status of the PV inverter.
+            set_power_result_receiver: Receiver for the power assignment result.
         """
+        BackgroundService.__init__(
+            self, name=f"PVInverterStatusTracker({component_id})"
+        )
         self._component_id = component_id
         self._max_data_age = max_data_age
         self._status_sender = status_sender
@@ -76,69 +73,35 @@ class EVChargerStatusTracker(ComponentStatusTracker, BackgroundService):
             max_duration=max_blocking_duration,
         )
 
-        BackgroundService.__init__(self, name=f"EVChargerStatusTracker({component_id})")
-
     @override
     def start(self) -> None:
         """Start the status tracker."""
         self._tasks.add(asyncio.create_task(self._run_forever()))
 
-    def _is_working(self, ev_data: EVChargerData) -> bool:
-        """Return whether the given data indicates that the component is working."""
-        return ev_data.cable_state in (
-            EVChargerCableState.EV_PLUGGED,
-            EVChargerCableState.EV_LOCKED,
-        ) and ev_data.component_state in (
-            EVChargerComponentState.READY,
-            EVChargerComponentState.CHARGING,
-            EVChargerComponentState.DISCHARGING,
+    def _is_working(self, pv_data: InverterData) -> bool:
+        """Return whether the given data indicates that the PV inverter is working."""
+        return pv_data.component_state in (
+            InverterComponentState.DISCHARGING,
+            InverterComponentState.CHARGING,
+            InverterComponentState.IDLE,
+            InverterComponentState.STANDBY,
         )
 
-    def _is_stale(self, ev_data: EVChargerData) -> bool:
-        """Return whether the given data is stale."""
-        now = datetime.now(tz=timezone.utc)
-        stale = now - ev_data.timestamp > self._max_data_age
-        return stale
-
     async def _run_forever(self) -> None:
-        """Run the status tracker forever."""
         while True:
             try:
                 await self._run()
             except Exception:  # pylint: disable=broad-except
                 _logger.exception(
-                    "Restarting after exception in EVChargerStatusTracker.run()"
+                    "Restarting after exception in PVInverterStatusTracker.run()"
                 )
                 await asyncio.sleep(1.0)
 
-    def _handle_ev_data(self, ev_data: EVChargerData) -> ComponentStatusEnum:
-        """Handle new EV charger data."""
-        if self._is_stale(ev_data):
-            if self._last_status == ComponentStatusEnum.WORKING:
-                _logger.warning(
-                    "EV charger %s data is stale. Last timestamp: %s",
-                    self._component_id,
-                    ev_data.timestamp,
-                )
-            return ComponentStatusEnum.NOT_WORKING
-
-        if self._is_working(ev_data):
-            if self._last_status == ComponentStatusEnum.NOT_WORKING:
-                _logger.warning(
-                    "EV charger %s is in WORKING state.",
-                    self._component_id,
-                )
-            return ComponentStatusEnum.WORKING
-
-        if self._last_status == ComponentStatusEnum.WORKING:
-            _logger.warning(
-                "EV charger %s is in NOT_WORKING state. "
-                "Cable state: %s, component state: %s",
-                self._component_id,
-                ev_data.cable_state,
-                ev_data.component_state,
-            )
-        return ComponentStatusEnum.NOT_WORKING
+    def _is_stale(self, pv_data: InverterData) -> bool:
+        """Return whether the given data is stale."""
+        now = datetime.now(tz=timezone.utc)
+        stale = now - pv_data.timestamp > self._max_data_age
+        return stale
 
     def _handle_set_power_result(
         self, set_power_result: SetPowerResult
@@ -150,16 +113,43 @@ class EVChargerStatusTracker(ComponentStatusTracker, BackgroundService):
         self._blocking_status.block()
         if self._last_status == ComponentStatusEnum.WORKING:
             _logger.warning(
-                "EV charger %s is in UNCERTAIN state. Set power result: %s",
+                "PV inverter %s is in UNCERTAIN state. Set power result: %s",
                 self._component_id,
                 set_power_result,
             )
         return ComponentStatusEnum.UNCERTAIN
 
+    def _handle_pv_inverter_data(self, pv_data: InverterData) -> ComponentStatusEnum:
+        """Handle new PV inverter data."""
+        if self._is_stale(pv_data):
+            if self._last_status == ComponentStatusEnum.WORKING:
+                _logger.warning(
+                    "PV inverter %s data is stale. Last timestamp: %s",
+                    self._component_id,
+                    pv_data.timestamp,
+                )
+            return ComponentStatusEnum.NOT_WORKING
+
+        if self._is_working(pv_data):
+            if self._last_status == ComponentStatusEnum.NOT_WORKING:
+                _logger.warning(
+                    "PV inverter %s is in WORKING state.",
+                    self._component_id,
+                )
+            return ComponentStatusEnum.WORKING
+
+        if self._last_status == ComponentStatusEnum.WORKING:
+            _logger.warning(
+                "PV inverter %s is in NOT_WORKING state.  Component state: %s",
+                self._component_id,
+                pv_data.component_state,
+            )
+        return ComponentStatusEnum.NOT_WORKING
+
     async def _run(self) -> None:
         """Run the status tracker."""
         api_client = connection_manager.get().api_client
-        ev_data_rx = await api_client.ev_charger_data(self._component_id)
+        pv_data_rx = await api_client.inverter_data(self._component_id)
         set_power_result_rx = self._set_power_result_receiver
         missing_data_timer = Timer(self._max_data_age, SkipMissedAndDrift())
 
@@ -169,17 +159,18 @@ class EVChargerStatusTracker(ComponentStatusTracker, BackgroundService):
         )
 
         async for selected in select(
-            ev_data_rx, set_power_result_rx, missing_data_timer
+            pv_data_rx, set_power_result_rx, missing_data_timer
         ):
             new_status = ComponentStatusEnum.NOT_WORKING
-            if selected_from(selected, ev_data_rx):
+            if selected_from(selected, pv_data_rx):
                 missing_data_timer.reset()
-                new_status = self._handle_ev_data(selected.message)
+                new_status = self._handle_pv_inverter_data(selected.message)
             elif selected_from(selected, set_power_result_rx):
                 new_status = self._handle_set_power_result(selected.message)
             elif selected_from(selected, missing_data_timer):
                 _logger.warning(
-                    "No EV charger %s data received for %s. Setting status to NOT_WORKING.",
+                    "No PV inverter %s data received for %s. "
+                    "Setting status to NOT_WORKING.",
                     self._component_id,
                     self._max_data_age,
                 )
@@ -193,7 +184,7 @@ class EVChargerStatusTracker(ComponentStatusTracker, BackgroundService):
 
             if new_status != self._last_status:
                 _logger.info(
-                    "EV charger %s status changed from %s to %s",
+                    "PV inverter %s status changed from %s to %s",
                     self._component_id,
                     self._last_status,
                     new_status,
