+++ src/frequenz/sdk/microgrid/_power_distributing/result.py
