+++ src/frequenz/sdk/timeseries/battery_pool/_methods.py
@@ -14,7 +14,7 @@ from frequenz.channels import Broadcast, Receiver
 
 from ..._internal._asyncio import cancel_and_await
 from ..._internal._constants import RECEIVER_MAX_SIZE, WAIT_FOR_COMPONENT_DATA_SEC
-from ...actor.power_distributing._component_managers._battery_manager import (
+from ...microgrid._power_distributing._component_managers._battery_manager import (
     _get_battery_inverter_mappings,
 )
 from ._component_metric_fetcher import (
