+++ src/frequenz/sdk/timeseries/ev_charger_pool/_system_bounds_tracker.py
@@ -12,8 +12,8 @@ from frequenz.channels import Receiver, Sender, merge, select, selected_from
 from frequenz.client.microgrid import EVChargerData
 
 from ...actor import BackgroundService
-from ...actor.power_distributing._component_status import ComponentPoolStatus
 from ...microgrid import connection_manager
+from ...microgrid._power_distributing._component_status import ComponentPoolStatus
 from .. import Power
 from .._base_types import Bounds, SystemBounds
